change renderer to be imported as a module.
